Split functionality to different files
